requirements_ddh_dev.txt
psutil
setproctitle
pyside6
redis
numpy
pyqtgraph
python-dateutil
pandas
GitPython
toml
tzlocal
pyserial
pynmea2
gpiozero
requests
strip.pysondb
bleak
humanize
boto3
h5py


# don't install during dev but "attach" in pycharm
# gps@git+https://github.com/LowellInstruments/gps.git
# ble@git+https://github.com/LowellInstruments/ble.git


# for BRT tool
# simplepyble~=0.7.3
# lsb@git+https://github.com/LowellInstruments/lsb.git
